Fixed room type creation

+++ src/main/java/com/example/HotelSPP/entity/RoomType.java
@@ -61,7 +61,7 @@ public class RoomType {
         this.amount = quantity;
     }
 
-    public double getPrice() {
+    public float getPrice() {
         return price;
     }
